requirements.txt
Babel==2.15.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-Security==3.0.0
Flask-WTF==1.2.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pymongo==4.7.3
python-dotenv==1.0.1
speaklater==1.3
Werkzeug==3.0.3
WTForms==3.1.2